More fixes to dodgy signatures

+++ src/transformers/models/funnel/modeling_tf_funnel.py
@@ -1522,7 +1522,7 @@ class TFFunnelForMultipleChoice(TFFunnelPreTrainedModel, TFMultipleChoiceLoss):
     def serving_signature(self) -> Dict[str, tf.TypeSpec]:
         return {
             "input_ids": tf.TensorSpec((None, None), tf.int32, name="input_ids"),
-            "attention_mask": tf.TensorSpec((None, None), tf.float32, name="attention_mask"),
+            "attention_mask": tf.TensorSpec((None, None), tf.int32, name="attention_mask"),
             "token_type_ids": tf.TensorSpec((None, None), tf.int32, name="token_type_ids"),
         }
